+++ src/components/Animation/BsRipple.ts
@@ -1,17 +1,19 @@
 import type { ComponentOptionsMixin, ComputedOptions, EmitsOptions, MethodOptions } from 'vue';
-import { defineComponent, h, ref, watch } from 'vue';
-import { cssPrefix, useGenerateId } from '../../mixins/CommonApi';
+import { computed, defineComponent, ref, watch } from 'vue';
+import { cssPrefix } from '../../mixins/CommonApi';
 import { booleanProp } from '../../mixins/CommonProps';
-import type { IRippleEvent, TBsRipple, TRecord, TRippleOptionProps } from '../../types';
-import Helper from '../../utils/Helper';
+import type { TBsRipple, TRecord, TRippleOptionProps } from '../../types';
 import { baseTagProps } from '../Card/mixins/cardProps';
-import { useCreateRipple } from './mixins/rippleApi';
+import { startRipple, type TRippleData, useRenderRipples } from './mixins/rippleApi';
 
 export default defineComponent<TBsRipple, TRecord, TRecord, ComputedOptions, MethodOptions, ComponentOptionsMixin, ComponentOptionsMixin, EmitsOptions>({
     name: 'BsRipple',
     props: {
         ...baseTagProps,
-        active: booleanProp,
+        active: {
+            type: [Boolean, Event],
+            default: undefined,
+        },
         centered: booleanProp,
         disabled: booleanProp,
     },
@@ -22,63 +24,47 @@ export default defineComponent<TBsRipple, TRecord, TRecord, ComputedOptions, Met
         'update:active',
     ],
     setup(props, {emit, slots}) {
-        const cmpProps = props as Readonly<TRippleOptionProps>;
-        const cmpId = useGenerateId();
-        const touchTimeout = ref<number>();
-        const startRipple = (event: IRippleEvent): void => {
-            emit('update:active', true);
-            useCreateRipple(event, cmpProps.centered);
-            Helper.defer(() => {
+        const thisProps = props as Readonly<TRippleOptionProps>;
+        const ripples = ref<TRippleData[]>([]);
+        const element = ref<HTMLElement | null>(null);
+        const touchTimeout = ref<number>(0);
+        const eventType = ref<string>('');
+        const disabled = ref(thisProps.disabled ?? false);
+        const centered = ref(thisProps.centered ?? false);
+        const classNames = computed<TRecord>(() => ({
+            [`${cssPrefix}ripple`]: true,
+            'ripple-off': disabled.value,
+        }));
+        const rippleClassNames = computed<TRecord>(() => ({
+            [`${cssPrefix}center`]: centered.value,
+        }));
+
+        watch(
+            () => thisProps.active,
+            (value) => {
+                const isBoolean = typeof value === 'boolean';
+                // @ts-ignore
+                const isEvent = value?.constructor.toString().match(/function (\w*)/)[1].toLowerCase() === 'mouseevent';
+
+                if (isBoolean && !disabled.value && value) {
+                    startRipple(
+                        element, ripples, eventType, disabled, centered,
+                        {type: 'mousedown'} as MouseEvent & TouchEvent
+                    );
+                } else if (isEvent) {
+                    startRipple(
+                        element, ripples, eventType, disabled, centered,
+                        value as MouseEvent & TouchEvent
+                    );
+                }
                 emit('update:active', false);
-            }, 100);
-        }
-
-        watch(
-            () => cmpProps.active,
-            (value) => {
-                if (value === true && !cmpProps.disabled) {
-                    const event = {target: document.getElementById(cmpId)} as IRippleEvent;
-                    useCreateRipple(event, cmpProps.centered);
-                    Helper.defer(() => {
-                        emit('update:active', false);
-                    }, 100);
-                }
-            }
-        );
-        watch(
-            () => cmpProps.disabled,
-            (value) => {
-                if (value === true) {
-                    const target = document.getElementById(cmpId);
-
-                    if (target) {
-                        const ripple = target.getElementsByClassName(`${cssPrefix}ripple-animation`)[0];
-                        if (ripple) {
-                            ripple.remove();
-                        }
-                    }
-                }
             }
         );
 
-        return () => {
-            return h(props.tag as string, {
-                id: cmpId,
-                class: `${cssPrefix}ripple`,
-                onClickPassive: (event: IRippleEvent) => {
-                    !props.disabled && startRipple(event);
-                },
-                onTouchstartPassive: (event: IRippleEvent) => {
-                    if (!props.disabled) {
-                        touchTimeout.value = window.setTimeout(() => {
-                            startRipple(event);
-                        }, 100);
-                    }
-                },
-                onTouchmovePassive: () => {
-                    window.clearTimeout(touchTimeout.value);
-                }
-            }, slots.default && slots.default())
-        }
-    }
+        return () =>
+            useRenderRipples(
+                slots, element, ripples, classNames, rippleClassNames,
+                eventType, disabled, centered, touchTimeout, thisProps.tag,
+            );
+    },
 });