+++ src/components/Menu/BsDropdownMenu.ts
@@ -1,5 +1,5 @@
 import type { ComponentOptionsMixin, ComponentPublicInstance, ComputedOptions, EmitsOptions, MethodOptions } from 'vue';
-import { defineComponent, ref } from 'vue';
+import { defineComponent, ref, watch } from 'vue';
 import { booleanProp, booleanTrueProp, validStringOrNumberProp, whiteColorProp } from '../../mixins/CommonProps';
 import type { TBsDropdownMenu, TDropdownMenuOptionProps, TRecord } from '../../types';
 import { popoverDefaultTransitionProp, popoverPlacementProp } from '../Popover/mixins/popoverProps';
@@ -34,6 +34,13 @@ export default defineComponent<TBsDropdownMenu, TRecord, TRecord, ComputedOption
         const popupMenu = ref<ComponentPublicInstance | null>(null);
         const isActive = ref(<boolean>thisProps.open);
         const timer = ref<number>();
+
+        watch(
+            () => thisProps.open,
+            (value) => {
+                isActive.value = <boolean>value;
+            }
+        );
 
         return () =>
             useRenderDropdownMenu(props, slots, emit, activator, popupMenu, timer, isActive)